Added a new file for get_location, fixed syntax in weather_info, and removed decimals in forecast data
