+++ weather.py
@@ -1,26 +1,13 @@
-'''Contains the function get_location(ZIPCODE) and get_weather(ZIPCODE).'''
+'''Contains the function and get_weather(ZIPCODE) which gets forecast data.'''
 import os
 import requests
 from dotenv import load_dotenv, find_dotenv
-#from date import convert
 from icon import get_icon
 from weather_info import pull_weather
+from locate import get_location
 
 load_dotenv(find_dotenv())
 WEATHER_KEY = os.getenv('WEATHER_KEY')
-
-
-def get_location(zipcode):
-    '''Uses another OpenWeather API to get latitude and longitude'''
-    url = "https://api.openweathermap.org/data/2.5/weather?zip={ZIPCODE}&appid={KEY}".format(
-        ZIPCODE=zipcode, KEY=WEATHER_KEY)
-    data = requests.get(url)
-    data = data.json()
-    lon = str(data["coord"]["lon"])
-    lat = str(data["coord"]["lat"])
-    town = str(data["name"])
-    return lat, lon, town
-
 
 def get_weather(zipcode):
     '''Will fetch the weather data using the latitude and longitude of a location'''
@@ -34,6 +21,7 @@ def get_weather(zipcode):
     data = data.json()
 
     current_temp = str(data["current"]["temp"])
+    current_temp = str(round(float(current_temp)))
     current_weather = str(data["current"]["weather"][0]["main"])
     icon_curr = str(data["current"]["weather"][0]["icon"])
     icon_curr = get_icon(icon_curr)
